+++ src/main/java/com/gabozago/hack/dto/review/ReviewPostDto.java
@@ -3,6 +3,7 @@ package com.gabozago.hack.dto.review;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 
 @Data
@@ -11,11 +12,11 @@ public class ReviewPostDto {
     private Long placeId;
     private ArrayList<String> keywords;
     private String content;
-    private Long rate;
+    private BigDecimal rate;
     private String image;
 
     @Builder
-    public ReviewPostDto(Long userId, Long placeId, ArrayList<String> keywords, String content, Long rate, String image){
+    public ReviewPostDto(Long userId, Long placeId, ArrayList<String> keywords, String content, BigDecimal rate, String image){
         this.userId = userId;
         this.placeId = placeId;
         this.keywords = keywords;
